feat(ia-test-reward): test neural network setup

+++ src/ai/behaviors/Behavior.cpp
@@ -7,5 +7,5 @@
 
 #include "Behavior.hpp"
 
-Behavior::Behavior(int probability, std::function<void()> act, std::string name) : probability(probability), act(act), name(name){};
+Behavior::Behavior(double probability, std::function<void()> act, std::string name) : probability(probability), act(act), name(name){};
 Behavior::~Behavior(){};
